% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_clone_fraction_from_ccf.R
\name{adjust_clone_fraction_from_ccf}
\alias{adjust_clone_fraction_from_ccf}
\title{adjust_clone_fraction_from_ccf}
\usage{
adjust_clone_fraction_from_ccf(clone_fraction, tree, node)
}
\arguments{
\item{clone_fraction}{clone CCF}

\item{tree}{tree}

\item{node}{node}
}
\value{
adjusted clone CCF
}
\description{
\code{adjust_clone_fraction_from_ccf} returns adjusted clone CCF based on tree
}
\keyword{internal}
