% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_ccf_2r.R
\docType{data}
\name{example_ccf_2r}
\alias{example_ccf_2r}
\title{Example data for MAPSCE package}
\format{
3 tibbles and 2 vectors.
}
\usage{
data(example_data)
}
\description{
A dataset containing example patient, copy number, cluster CCF, mutational CCF and tree.
}
\details{
\itemize{
\item example_patient_ID. example patient
\item example_cn. example copy number
\item example_ccf. example cluster CCF
\item example_mutational_ccf. example mutational CCF.
\item example_tree. example tree.
\item example_patient_ID_2r. example patient for 2 regions
\item example_cn_2r. example copy number for 2 regions
\item example_ccf_2r. example cluster CCF for 2 regions
\item example_tree_2r. example tree for 2 regions
}
}
\keyword{datasets}
