% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_mutational_ccf.R
\docType{data}
\name{example_mutational_ccf}
\alias{example_mutational_ccf}
\title{Example data for MAPSCE package}
\format{
3 tibbles and 2 vectors.
}
\usage{
data(example_data)
}
\description{
A dataset containing example patient, copy number, cluster CCF, mutational CCF and tree.
}
\details{
\itemize{
\item example_patient_ID. example patient
\item example_cn. example copy number
\item example_ccf. example cluster CCF
\item example_mutational_ccf. example mutational CCF.
\item example_tree. example tree.
}
}
\keyword{datasets}
