% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clone_fraction_from_ccf.R
\name{get_clone_fraction_from_ccf}
\alias{get_clone_fraction_from_ccf}
\title{get_clone_fraction_from_ccf}
\usage{
get_clone_fraction_from_ccf(ccf, tree)
}
\arguments{
\item{ccf}{ccf}

\item{tree}{tree}
}
\value{
a matrix
}
\description{
\code{get_clone_fraction_from_ccf} returns clone CCF from cluster CCF
}
\keyword{internal}
