% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_consensus.R
\name{get_consensus}
\alias{get_consensus}
\title{get_consensus}
\usage{
get_consensus(data, this_patient, consensus = TRUE, consensus.only = F)
}
\arguments{
\item{data}{summarised mapping results from mapsce function}

\item{this_patient}{patient ID}

\item{consensus}{printing of the consensus}

\item{consensus.only}{limiting output to just the consensus}
}
\value{
returns a matrix with consensus copy number states for every clone in a tree, labeling no consensus as NAs
}
\description{
\code{get_consensus} finds a consensus copy number results for a particular mapping result
}
