% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nested_set_from_tree.R
\name{get_nested_set_from_tree}
\alias{get_nested_set_from_tree}
\title{get_nested_set_from_tree}
\usage{
get_nested_set_from_tree(tree)
}
\arguments{
\item{tree}{tree}
}
\value{
a matrix
}
\description{
\code{get_nested_set_from_tree} returns a nested set from a tree
}
