% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_tree.R
\name{index_tree}
\alias{index_tree}
\title{index_tree}
\usage{
index_tree(tree)
}
\arguments{
\item{tree}{patient tree}
}
\value{
tibble with indices
}
\description{
\code{index_tree} indexes all branches in a given tree
}
\keyword{internal}
