% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_mapsce.R
\name{interpret_mapsce}
\alias{interpret_mapsce}
\title{interpret_mapsce}
\usage{
interpret_mapsce(
  data,
  tree,
  min_diff = 0.4,
  consensus_threshold = 0.1,
  format = "tibble",
  graph = F
)
}
\arguments{
\item{data}{a tibble with summarised mapping results, the output of using mapsce function}

\item{tree}{a matrix listing all the branches in the tree, where the first column is the ancestral node and the second column is the descendant clone. By definition, the root node will only be present in the first column. The clone IDs must correspond to the cluster IDs in cluster_ccf and mutation_ccf.}

\item{min_diff}{a numeric vector, minimum difference between before and after to consider clonality. default at 0.4}

\item{consensus_threshold}{a numeric vector, the allowed threshold between the values to be considered "in agreement/consensus". default at 0.1}

\item{format}{a character vector, "list" or "tibble" output format. default is tibble}

\item{graph}{logical, graphical output}
}
\value{
returns interpretation of MAPSCE in the form of vector
}
\description{
\code{interpret_mapsce} is a wrapper function for MAPSCE for interpretation of results for a particular mapping result
}
\examples{
data(example_data)

mapsce_result <- mapsce(
example_cn,
example_ccf,
example_mutational_ccf,
example_tree)

interpret_mapsce(mapsce_result, example_tree)
# returns an interpretation vector for this example mapsce result


}
