% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPSCE.R
\name{mapsce}
\alias{mapsce}
\title{mapsce}
\usage{
mapsce(
  patient,
  copy_number,
  cluster_ccf,
  mutation_ccf,
  tree,
  bootstraps = 100,
  print_raw_matrix = "no",
  print_duration = "yes"
)
}
\arguments{
\item{patient}{patient ID}

\item{copy_number}{observed copy numbers}

\item{cluster_ccf}{CCF matrix}

\item{mutation_ccf}{mutational CCF matrix}

\item{tree}{tree matrix}

\item{bootstraps}{number of bootstraps to run}

\item{print_raw_matrix}{logical}

\item{print_duration}{printing time running}
}
\value{
Returns tibble with summarised bootstrapped mapping results
}
\description{
\code{mapsce} returns branch test results.
}
