% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapsce2r.R
\name{mapsce2r}
\alias{mapsce2r}
\title{mapsce2r}
\usage{
mapsce2r(
  copy_number,
  cluster_ccf,
  tree,
  print_raw_matrix = "no",
  print_duration = "yes"
)
}
\arguments{
\item{copy_number}{observed copy numbers}

\item{cluster_ccf}{CCF matrix}

\item{tree}{matrix with tree topology}

\item{print_raw_matrix}{printing of raw results}

\item{print_duration}{printing of the time taken to run}
}
\value{
a tibble with column names:
\itemize{
\item branch. branch ID
\item before. copy number state before
\item after. copy number state after
\item rss. residual square of sums
\item btn. how many bootstrapped BICs were better than the null's BI
\item nregions. number of regions
\item nclones. number of clones or branches.
\item null. whether the branch is a trunk or not.
\item bic.  bic calculated from mean RSS
\item bf. the strength of evidence of bayes factor comparison
\item evid. 1 for good results based on the Bayes Factor comparison
\item index. ordering of results.
}
}
\description{
\code{mapsce2r} returns a matrix with a branch test for every single branch of a tree
}
\examples{
data(example_data)
mapsce2r(example_cn_2r, example_ccf_2r, example_tree_2r)
#returns a tibble with branch 7 as the best result

}
