% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_pyclone_ccf.R
\name{reduce_pyclone_ccf}
\alias{reduce_pyclone_ccf}
\title{reduce_pyclone_ccf}
\usage{
reduce_pyclone_ccf(ccf, tree)
}
\arguments{
\item{ccf}{cluster ccf}

\item{tree}{tree}
}
\value{
reduced cluster ccf
}
\description{
\code{reduce_pyclone_ccf} returns a reduced cluster ccf
}
\keyword{internal}
