% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_consensus.R
\name{get_consensus}
\alias{get_consensus}
\title{get_consensus}
\usage{
get_consensus(data, this_patient, consensus = TRUE, consensus.only = F)
}
\arguments{
\item{data}{summarised mapping results}

\item{this_patient}{patient ID}

\item{consensus}{logical}

\item{consensus.only}{logical}
}
\value{
returns a consensus matrix
}
\description{
\code{get_consensus} finds a consensus copy number results for a particular mapping result
}
