% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_consensus.R
\name{get_consensus}
\alias{get_consensus}
\title{get_consensus}
\usage{
get_consensus(data, tree, consensus = TRUE, consensus.only = F)
}
\arguments{
\item{data}{a tibble with summarised mapping results, the output of using mapsce function}

\item{tree}{a matrix listing all the branches in the tree, where the first column is the ancestral node and the second column is the descendant clone. By definition, the root node will only be present in the first column. The clone IDs must correspond to the cluster IDs in cluster_ccf and mutation_ccf.}

\item{consensus}{printing of the consensus}

\item{consensus.only}{limiting output to just the consensus}
}
\value{
returns a matrix with consensus copy number states for every clone in a tree, labeling no consensus as NAs
}
\description{
\code{get_consensus} finds a consensus copy number results for a particular mapping result
}
\examples{
data(example_data)

mapsce_result <- mapsce(
example_cn,
example_ccf,
example_mutational_ccf,
example_tree)

get_consensus(mapsce_result, example_tree)
# returns a consensus matrix for this example mapsce result


}
