% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPSCE.R
\name{mapsce}
\alias{mapsce}
\title{mapsce}
\usage{
mapsce(
  patient,
  copy_number,
  cluster_ccf,
  mutation_ccf,
  tree,
  bootstraps = 100,
  print_raw_matrix = "no",
  print_duration = "yes"
)
}
\arguments{
\item{patient}{patient ID}

\item{copy_number}{observed copy numbers}

\item{cluster_ccf}{CCF matrix}

\item{mutation_ccf}{mutational CCF matrix}

\item{tree}{matrix with tree topology}

\item{bootstraps}{number of bootstraps}

\item{print_raw_matrix}{printing of raw results}

\item{print_duration}{printing of the time taken to run}
}
\value{
a tibble
\itemize{
\item branch. branch ID
\item before. copy number state before
\item after. copy number state after
\item rss. average residual square of sums of all bootstraps
\item btn. how many bootstrapped BICs were better than the null's BIC
\item number_of_regions. number of regions
\item nclones. number of clones or branches.
\item null. whether the branch is a trunk or not.
\item bic. average bic for all bootstraps calculated from mean RSS
\item bf. the strength of evidence of bayes factor comparison
\item evid. 1 for good results based on the Bayes Factor comparison and btn.
}
}
\description{
\code{mapsce} returns a matrix with a branch test for every single branch of a tree
}
\examples{
data(example_data.RData)
mapsce(example_patient_ID, example_cn, example_ccf, example_mutational_ccf, example_tree)
#returns a tibble with branch 7 as the best result

}
