% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPSCE.R
\name{mapsce}
\alias{mapsce}
\title{mapsce}
\usage{
mapsce(
  patient,
  copy_number,
  cluster_ccf,
  mutation_ccf,
  tree,
  bootstraps = 100,
  print_raw_matrix = "no",
  print_duration = "yes"
)
}
\arguments{
\item{patient}{patient ID}

\item{copy_number}{observed copy numbers}

\item{cluster_ccf}{CCF matrix}

\item{mutation_ccf}{mutational CCF matrix}

\item{tree}{matrix with tree topology}

\item{bootstraps}{number of bootstraps}

\item{print_raw_matrix}{printing of raw results}

\item{print_duration}{printing of the time taken to run}
}
\value{
a matrix with branch ID, copy number states before and after, residual sum of squares, number of clones and regions, whether the branch was the trunk the bic of the branch and the bayes factor comparison.
}
\description{
\code{mapsce} returns a matrix with a branch test for every single branch of a tree
}
