% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPSCE.R
\name{mapsce}
\alias{mapsce}
\title{mapsce}
\usage{
mapsce(
  copy_number,
  cluster_ccf,
  mutation_ccf,
  tree,
  bootstraps = 100,
  print_raw_matrix = F,
  print_duration = T,
  force_bootstrap = F,
  clone_ccf = F
)
}
\arguments{
\item{copy_number}{a numeric vector with copy number values ordered by sample}

\item{cluster_ccf}{a matrix with mean CCF values for clones in rows and samples (same order as in copy_number) in columns}

\item{mutation_ccf}{a tibble or data frame where the first columns contain the mutational CCF values for the  samples (same order as in copy_number) in columns. Tw of the columns in this object need to be "PycloneCluster" and "CleanCluster", the former being the clone assignment and the latter being the PyClone filter (1 for valid clusters, 0 for clusters to be ignored)}

\item{tree}{a matrix listing all the branches in the tree, where the first column is the ancestral node and the second column is the descendant clone. By definition, the root node will only be present in the first column. The clone IDs must correspond to the cluster IDs in cluster_ccf and mutation_ccf.}

\item{bootstraps}{number of bootstraps for reclustering of the CCF}

\item{print_raw_matrix}{logical, printing of raw results}

\item{print_duration}{logical, printing of the time taken to run}

\item{force_bootstrap}{logical, forcing mapsce to run bootstrapping for patients with 2 regions only}

\item{clone_ccf}{logical, forcing mapsce to use clone CCF instead of just cluster CCF}
}
\value{
a tibble with column names:
\itemize{
\item branch. branch ID
\item before. copy number state before
\item after. copy number state after
\item rss. average residual square of sums of all bootstraps
\item btn. how many bootstrapped BICs were better than the null's BIC
\item nregions. number of regions
\item nclones. number of clones or branches.
\item null. whether the branch is a trunk or not.
\item bic. average bic for all bootstraps calculated from mean RSS
\item bf. the strength of evidence of bayes factor comparison
\item evid. 1 for good results based on the Bayes Factor comparison and btn
\item index. ordering of results.
}
}
\description{
\code{mapsce} returns a matrix with a branch test for every single branch of a tree
}
\examples{
data(example_data)
mapsce(example_cn, example_ccf, example_mutational_ccf, example_tree)
#returns a tibble with branch 7 as the best result

}
