% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapsce2r.R
\name{mapsce2r}
\alias{mapsce2r}
\title{mapsce2r}
\usage{
mapsce2r(
  copy_number,
  cluster_ccf,
  tree,
  print_raw_matrix = F,
  print_duration = T
)
}
\arguments{
\item{copy_number}{a numeric vector with copy number values ordered by sample}

\item{cluster_ccf}{a matrix with mean CCF values for clones in rows and samples (same order as in copy_number) in columns}

\item{tree}{a matrix listing all the branches in the tree, where the first column is the ancestral node and the second column is the descendant clone. By definition, the root node will only be present in the first column. The clone IDs must correspond to the cluster IDs in cluster_ccf and mutation_ccf.}

\item{print_raw_matrix}{logical, printing of raw results}

\item{print_duration}{logical, printing of the time taken to run}
}
\value{
a tibble with column names:
\itemize{
\item branch. branch ID
\item before. copy number state before
\item after. copy number state after
\item rss. residual square of sums
\item btn. how many bootstrapped BICs were better than the null's BI
\item nregions. number of regions
\item nclones. number of clones or branches.
\item null. whether the branch is a trunk or not.
\item bic.  bic calculated from mean RSS
\item bf. the strength of evidence of bayes factor comparison
\item evid. 1 for good results based on the Bayes Factor comparison
\item index. ordering of results.
}
}
\description{
\code{mapsce2r} returns a matrix with a branch test for every single branch of a tree
}
\examples{
data(example_data)
mapsce2r(example_cn_2r, example_ccf_2r, example_tree_2r)
#returns a tibble with branch 2 as the best result

}
